+++ app/src/client_audio.h
@@ -1,3 +1,6 @@
+#ifndef SC_CLIENT_AUDIO_H
+#define SC_CLIENT_AUDIO_H
+
 #include "util/net.h"
 
 struct sc_microphone_params {
@@ -10,3 +13,5 @@ sc_microphone_list_audio_sources(void);
 
 int
 sc_microphone_run(void *data);
+
+#endif
